- Merge coroutine into core. - Merge cocos_coroutine into cocos.
