+++ Android.mk
@@ -8,9 +8,7 @@ $(call import-module, src/ee/app_lovin)
 $(call import-module, src/ee/apps_flyer)
 $(call import-module, src/ee/campaign_receiver)
 $(call import-module, src/ee/cocos)
-$(call import-module, src/ee/cocos_coroutine)
 $(call import-module, src/ee/core)
-$(call import-module, src/ee/coroutine)
 $(call import-module, src/ee/crashlytics)
 $(call import-module, src/ee/facebook)
 $(call import-module, src/ee/facebook_ads)
