+++ CMakeLists.txt
@@ -8,9 +8,7 @@ add_subdirectory(src/ee/app_lovin)
 add_subdirectory(src/ee/apps_flyer)
 add_subdirectory(src/ee/campaign_receiver)
 add_subdirectory(src/ee/cocos)
-add_subdirectory(src/ee/cocos_coroutine)
 add_subdirectory(src/ee/core)
-add_subdirectory(src/ee/coroutine)
 add_subdirectory(src/ee/crashlytics)
 add_subdirectory(src/ee/facebook)
 add_subdirectory(src/ee/facebook_ads)
