+++ src/ee/Core.hpp
@@ -11,10 +11,14 @@
 
 #ifdef __cplusplus
 #include "ee/CoreFwd.hpp"
+#include "ee/core/AsyncMutex.hpp"
 #include "ee/core/IMessageBridge.hpp"
 #include "ee/core/IService.hpp"
+#include "ee/core/LambdaAwaiter.hpp"
 #include "ee/core/LogLevel.hpp"
 #include "ee/core/Logger.hpp"
+#include "ee/core/MakeAwaiter.hpp"
+#include "ee/core/NoAwait.hpp"
 #include "ee/core/ObserverBinder.hpp"
 #include "ee/core/ObserverHandle.hpp"
 #include "ee/core/ObserverManager.hpp"
@@ -23,9 +27,12 @@
 #include "ee/core/ScopeGuard.hpp"
 #include "ee/core/ServiceLocator.hpp"
 #include "ee/core/SpinLock.hpp"
+#include "ee/core/SwitchToUiThread.hpp"
+#include "ee/core/Task.hpp"
 #include "ee/core/Utils.hpp"
 #include "ee/core/VideoPlayer.hpp"
 #include "ee/core/VideoPlayerManager.hpp"
+#include "ee/core/WhenAll.hpp"
 #endif // __cplusplus
 
 #if defined(EE_X_ANDROID)
