+++ src/ee/ads/GuardedAdView.cpp
@@ -4,9 +4,9 @@
 
 #include <ee/core/ObserverHandle.hpp>
 #include <ee/core/SpinLock.hpp>
+#include <ee/core/SwitchToUiThread.hpp>
+#include <ee/core/Task.hpp>
 #include <ee/core/Utils.hpp>
-#include <ee/coroutine/SwitchToUiThread.hpp>
-#include <ee/coroutine/Task.hpp>
 
 namespace ee {
 namespace ads {
