+++ src/ee/ads/IRewardedAd.hpp
@@ -11,7 +11,6 @@
 
 #include <functional>
 
-#include <ee/CoroutineFwd.hpp>
 #include <ee/core/IObserverManager.hpp>
 
 #include "ee/AdsFwd.hpp"
