+++ src/ee/ads/MultiAdView.cpp
@@ -9,7 +9,7 @@
 #include "ee/ads/MultiAdView.hpp"
 
 #include <ee/core/ObserverHandle.hpp>
-#include <ee/coroutine/Task.hpp>
+#include <ee/core/Task.hpp>
 
 namespace ee {
 namespace ads {
