--- src/ee/cocos_coroutine/generate.sh
@@ -1,2 +1,0 @@
-SRC=$(printf "target_sources(\${PROJECT_NAME} PRIVATE\n$(find . -name "*.cpp" | sed "s/^/  \${CMAKE_CURRENT_LIST_DIR}\//")\n)\n");
-echo "$SRC" | cmp -s sourcelist.cmake - || echo "$SRC" > sourcelist.cmake
