+++ src/ee/core/Android.mk
@@ -5,7 +5,7 @@ include $(CLEAR_VARS)
 LOCAL_MODULE := ee_x_core
 LOCAL_ARM_MODE := arm
 
-LOCAL_CPP_FEATURES := exceptions
+LOCAL_CPP_FEATURES := exceptions coroutine-ts
 LOCAL_CPPFLAGS := -std=c++2a
 
 LOCAL_SRC_FILES := ${shell find $(LOCAL_PATH) -name "*.cpp" -print}
