--- src/ee/coroutine/Android.mk
@@ -1,12 +1,0 @@
-LOCAL_PATH := $(call my-dir)
-
-include $(CLEAR_VARS)
-
-LOCAL_MODULE := ee_x_coroutine
-LOCAL_CPP_FEATURES := exceptions coroutine-ts
-LOCAL_CPPFLAGS := -std=c++2a
-LOCAL_SRC_FILES := ${shell find $(LOCAL_PATH) -name "*.cpp" -print}
-
-LOCAL_STATIC_LIBRARIES := ee_x_core
-
-include $(BUILD_STATIC_LIBRARY)
